Try to make sure we handle EOF correctly
If EOF happens just after a \n, readlines() will not put an empty
string at the end of the list. Of course, this is also true if EOF
happens at the end of a line that doesn't terminate with \n. Handle
both of those cases by making sure that the last transaction, if any,
is written to output.

Because lines are written to output without \n, we need to make sure
every one has a \n at the end in __main__ when we write it out (rather
than just putting \n between lines).